Catch redirect() in root layout
