+++ packages/next/src/client/components/layout-router.tsx
@@ -1,9 +1,6 @@
 'use client'
 
-import type {
-  AppRouterInstance,
-  ChildSegmentMap,
-} from '../../shared/lib/app-router-context'
+import type { ChildSegmentMap } from '../../shared/lib/app-router-context'
 import type {
   FlightRouterState,
   FlightSegmentPath,
@@ -12,7 +9,7 @@ import type {
 import type { ErrorComponent } from './error-boundary'
 import { FocusAndScrollRef } from './router-reducer/router-reducer-types'
 
-import React, { useContext, useEffect, useMemo, use } from 'react'
+import React, { useContext, useMemo, use } from 'react'
 import ReactDOM from 'react-dom'
 import {
   CacheStates,
@@ -24,10 +21,9 @@ import { fetchServerResponse } from './router-reducer/fetch-server-response'
 import { createInfinitePromise } from './infinite-promise'
 import { ErrorBoundary } from './error-boundary'
 import { matchSegment } from './match-segments'
-import { useRouter } from './navigation'
 import { handleSmoothScroll } from '../../shared/lib/router/utils/handle-smooth-scroll'
-import { getURLFromRedirectError, isRedirectError } from './redirect'
 import { findHeadInCache } from './router-reducer/reducers/find-head-in-cache'
+import { RedirectBoundary } from './redirect-boundary'
 
 /**
  * Add refetch marker to router state at the point of the current layout segment.
@@ -420,55 +416,6 @@ function LoadingBoundary({
   }
 
   return <>{children}</>
-}
-
-interface RedirectBoundaryProps {
-  router: AppRouterInstance
-  children: React.ReactNode
-}
-
-function HandleRedirect({ redirect }: { redirect: string }) {
-  const router = useRouter()
-
-  useEffect(() => {
-    router.replace(redirect, {})
-  }, [redirect, router])
-  return null
-}
-
-class RedirectErrorBoundary extends React.Component<
-  RedirectBoundaryProps,
-  { redirect: string | null }
-> {
-  constructor(props: RedirectBoundaryProps) {
-    super(props)
-    this.state = { redirect: null }
-  }
-
-  static getDerivedStateFromError(error: any) {
-    if (isRedirectError(error)) {
-      const url = getURLFromRedirectError(error)
-      return { redirect: url }
-    }
-    // Re-throw if error is not for redirect
-    throw error
-  }
-
-  render() {
-    const redirect = this.state.redirect
-    if (redirect !== null) {
-      return <HandleRedirect redirect={redirect} />
-    }
-
-    return this.props.children
-  }
-}
-
-function RedirectBoundary({ children }: { children: React.ReactNode }) {
-  const router = useRouter()
-  return (
-    <RedirectErrorBoundary router={router}>{children}</RedirectErrorBoundary>
-  )
 }
 
 interface NotFoundBoundaryProps {
